requirements.txt
aiohttp == 3.8.3
google-cloud-bigquery == 3.11.0
apache-airflow-providers-google == 6.8.0
pyspark == 3.3.1
apache-airflow == 2.6.1
pandas == 1.5.2
lxml == 4.9.2
fsspec == 2023.6.0
gcsfs == 2023.5.0